+++ extensions/multi/CHANGELOG.md
@@ -1,10 +1,20 @@
 # Multi Changelog
 
+## [See session state for teammates] - 2024-04-08
+
+Requires Multi > 0.465.1
+
+- Call Teammate now shows if teammates are currently in a session
+  - If teammates are in a room, default action is to join the room instead of sending an invite.
+- Updated teammate availability icons
+
 ## [Updated metadata] - 2024-04-08
 
 Updated extension metadata on README.md and package.json
 
 ## [Initial Version] - 2024-04-04
+
+Requires Multi 0.460.1+
 
 Added initial set of commands: